+++ cfg4j-consul/src/main/java/consul/ConsulConfigurationSource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cfg4j.source.consul;
+package consul;
 
 import com.orbitz.consul.Consul;
 import com.orbitz.consul.KeyValueClient;
