+++ cfg4j-consul/src/test/java/consul/ConsulConfigurationSourceIntegrationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cfg4j.source.consul;
+package consul;
 
 import static org.assertj.core.api.Assertions.assertThat;
